Move BindingCode class to its own file
