+++ binding_generator/builtin_classes/constructor.py
@@ -1,8 +1,8 @@
 from textwrap import indent
 
+from common.binding_code import BindingCode
 from common.code_generator import CodeGenerator
-from format_utils import (BindingCode,
-                          format_arguments_array,
+from format_utils import (format_arguments_array,
                           format_parameter_const,
                           format_type_to_variant_enum,
                           should_generate_constructor)
