+++ binding_generator/builtin_classes/method.py
@@ -1,8 +1,8 @@
 from textwrap import indent
 
+from common.binding_code import BindingCode
 from common.code_generator import CodeGenerator
-from format_utils import (BindingCode,
-                          format_arguments_array,
+from format_utils import (format_arguments_array,
                           format_arguments_count,
                           format_parameter,
                           format_parameter_const,
