+++ binding_generator/common/scoped_enum.py
@@ -1,7 +1,7 @@
 from textwrap import indent
 
+from .binding_code import BindingCode
 from .code_generator import CodeGenerator
-from format_utils import BindingCode
 from json_types import *
