+++ binding_generator/format_utils.py
@@ -2,10 +2,8 @@
 Internal utilities for Godot types
 """
 
-from collections import defaultdict
 import re
-from textwrap import dedent, indent
-from typing import Sequence
+from textwrap import dedent
 
 from json_types import *
 
@@ -133,51 +131,6 @@ def should_generate_method(
         return method_name[4:] not in (member["name"] for member in members)
     else:
         return True
-
-
-class BindingCode:
-    """Object that contains the code necessary for each function binding"""
-    def __init__(self, prototype: str, implementation: str, **extras: str):
-        self.prototype = prototype
-        self.implementation = implementation
-        self.extras = extras
-
-    def __getitem__(self, key: str) -> str:
-        return self.extras.get(key, "")
-
-    def surround_prototype(self, prefix: str, suffix: str) -> None:
-        self.prototype = "\n".join([
-            prefix,
-            indent(self.prototype, "\t"),
-            suffix,
-        ])
-
-    def surround_implementation(self, prefix: str, suffix: str) -> None:
-        self.implementation = "\n".join([
-            prefix,
-            indent(self.implementation, "\t"),
-            suffix,
-        ])
-
-    def prepend_section_comment(self, comment: str):
-        if self.prototype:
-            self.prototype = f"// {comment}\n{self.prototype}"
-        if self.implementation:
-            self.implementation = f"// {comment}\n{self.implementation}"
-
-    @classmethod
-    def merge(cls, bindings: Sequence["BindingCode"], **kwargs: list[str]) -> "BindingCode":
-        extras: dict = defaultdict(list, **kwargs)
-        for b in bindings:
-            for k, v in b.extras.items():
-                extras[k].append(v)
-        for k, v in extras.items():
-            extras[k] = "\n".join(v)
-        return BindingCode(
-            "\n\n".join(b.prototype for b in bindings if b.prototype),
-            "\n\n".join(b.implementation for b in bindings if b.implementation),
-            **extras,
-        )
 
 
 ############################################################
